Add files via upload
Created a header section with a container class. 
Added search button to the header section.
Created a grid 7 columns by 2 rows to use as the navigation bar. This grid is also part of the header container class.
   -Each grid space will be a link to another page.
   -Row 1 will be an image link and Row 2 will be a text link. The image will correspond to the text on the row below it and they will both open a link to the same page.
Created a web page for each column link. 7 pages in total.
Downloaded the image icons from the AirBnb website for the 7 links that I created. The images are saved in a header icons folder which is located in the assets folder.