Add files via upload
19/06/24
Created a container with a light grey background to match the colour of the image in this section.
Inside the container:
Created a header 2 text for "Inspiration for future getaways" header
Created a grid 8 columns with 1 row for top navigation bar. For now these links will open a a new page.
Created a grid 2 columns with 4 rows for the main navigation section. For now these links will open a a new page.