Add files via upload
19/06/24
Changed the pastExperienceHeader class to headerText class. Previously used this to apply styling for all headers but will now break this class down and create new sub-classes. the new headerText will now only contain the properties that will be applied to all headers. The new sub-classes will contain new properties relevant to the new classes. I will break these down further if required.

This is to keep the code need and maintain good coding practice.

19/06/24
Have done the same as above to the numerous grids I have created so far. Have created a parent class called mainGrid with all the reusable code in this class. Will now create subclasses for any additional properties that need to be added for new grids. Will continue to create new subclaases if old code in those classes can be reused.

19/06/24
Created a grid 1 column with 7 rows. Used top row as a header and remaining rows as links to new pages. For now all links go to helpCenter page.