Add files via upload
20/06/24
Created a grid 1 column with 7 rows. Used top row as a header and remaining rows as links to new pages. For now all links go to helpCenter page. This was done twice. Once for the hosting section and one for the Airbnb section.
Also made the main.css file neater putting similar classes together.