Add files via upload
20/06/2024
downloaded globe icon and added to assets folder

23/06/2024
Created a flexbox. Inside the flexbox are 2 buttons which contain an image file and some text all of which are links to another page. For now all links are to holidayRental.html.
Underneath this are a series of buttons. Ive used &middot command and line breaks to make the page look the same as airbnb mobile page.