Add files via upload
Changing this section to closer resemble the mobile Airbnb website.

Image

The entire section should be a html link not just the text (as per my previous attempt). Also the background colour needs to be changed to the grey colour on the mobile airbnb site.

23/06/2024

Created a block level container and made the block the link. Text has been centred, underlined and bolded as per airbnb mobile website.