Add files via upload
Changing code to allow top search bar button to adjust size when resizing window. Currently search bar is fixed at an 800px size.

23/06/2024
Adding additional items on the top navigation bar and changing code to allow for horizontal scrolling. New links will open icons.html page for now.

24/06/2024
Changing the navigation header and search bar section to a fixed position has moved the rest of the content on the page up. This header now partially blocks part of this content. Have added an additional container of the same size behind the navigation header and search bar section to push the rest of the content back down so that everything can be viewed prior to the fixed position change.